requirements.txt
Django==1.11.4
Jinja2==2.9.6
django-rest-framework==0.1.0
djangorestframework==3.6.3
requests==2.18.3